+++ Tests/LinuxMain.swift
@@ -1,8 +1,0 @@
-import XCTest
-
-import BinaryCodableTests
-
-var tests = [XCTestCaseEntry]()
-tests += BinaryCodableTests.__allTests()
-
-XCTMain(tests)